extract `ModelInfoDialog` into a standalone component, integrate `handleDeleteBook` functionality for alignment memory deletion, and refine `EnhancedWordAligner` to use the extracted dialog component
